requirements.txt
langchain
streamlit
pinecone-client
langchain-pinecone
langchain-groq
python-dotenv
tqdm
pypdf
sentence-transformers
langchain-huggingface